update web-sdk to version 4.0.0 and new examples
