+++ custom-player-core/index.html
@@ -1,0 +1,71 @@
+<!DOCTYPE html>
+<html lang="en">
+  <head>
+    <meta charset="UTF-8" />
+    <meta http-equiv="X-UA-Compatible" content="IE=edge" />
+    <meta name="viewport" content="width=device-width, initial-scale=1.0" />
+    <title>Basic | Vindral Examples</title>
+    <!-- CSS Reset -->
+    <link
+      rel="stylesheet"
+      href="https://cdnjs.cloudflare.com/ajax/libs/normalize/8.0.1/normalize.css"
+    />
+
+    <!-- Milligram CSS -->
+    <link
+      rel="stylesheet"
+      href="https://cdnjs.cloudflare.com/ajax/libs/milligram/1.4.1/milligram.css"
+    />
+
+    <!-- Example CSS -->
+    <link
+      rel="stylesheet"
+      href="./index.css"
+    />
+  </head>
+  <body>
+    <main class="wrapper">
+      <nav class="navigation">
+        <section class="container">
+          <a class="button" href="../">Back to examples</a>
+        </section>
+      </nav>
+      <section class="container">
+        <h1>Custom Player Example</h1>
+        <p>
+          Here you can see a more advanced implementation of Vindral with a custom player interface.
+          In general we recommend using our <a href="../player/">Web Player</a> or the <a href="../embed-responsive">Embedded Player</a>.
+        </p>
+
+        <div id="playback-state">press start</div>
+
+        <div id="vindral-area" class="buffering">
+          <div id="vindral-container"></div>
+          <div id="activate-audio-button">
+            <svg xmlns="http://www.w3.org/2000/svg" viewBox="0 0 1024 896" height="100%" width="100%">
+              <path d="M128 384H0v256h128l256 192h64V192h-64L128 384zM864 416l-64-64-96 96-96-96-63 63.5 95 96.5-96 96 64 64 96-96 96 96 64-64-96-96L864 416z"/>
+            </svg>
+          </div>
+          <div id="vindral-thumbnail"></div>
+          <div class="buffering-overlay">
+            <div class="buffering-indicator"></div>
+          </div>
+
+          <div id="controls">
+            <button id="start-button">Start</button>
+            <button id="stop-button">Stop</button>
+            <button id="channel-button">New channel</button>
+          </div>
+        </div>
+
+      </section>
+
+      <section class="container" style="margin-top: 50px;">
+        <ul>
+          <li><a target="_blank" href="https://github.com/RealSprint/vindral-examples/tree/main/custom-player">Example source code</a></li>
+          <li><a target="_blank" href="https://docs.vindral.com/playout/websdk/">Web SDK Documentation</a></li>
+        </ul>
+      </section>
+    </main>
+  </body>
+</html>
