+++ index.html
@@ -58,8 +58,12 @@
               <td>The bare minimal code to get started with the Vindral Web SDK</td>
             </tr>
             <tr>
-              <td><a class="button" href="./custom-player/">Custom Player</a></td>
-              <td>Custom player GUI without using the bundled player</td>
+              <td><a class="button" href="./custom-player/">Custom Player - Web components</a></td>
+              <td>Custom player GUI using Vindral web components</td>
+            </tr>
+            <tr>
+              <td><a class="button" href="./custom-player-core/">Custom Player - Core SDK</a></td>
+              <td>Custom player GUI without using web components</td>
             </tr>
             <tr>
               <td><a class="button" href="./audio-only/">Audio-only</a></td>
