+++ player/index.html
@@ -28,6 +28,8 @@
         font-size: 1.2em;
       }
     </style>
+
+    <script type="module" async src="https://cdn.jsdelivr.net/npm/@vindral/web-sdk/vindral-player-component.js"></script>
   </head>
   <body>
     <main class="wrapper">
@@ -38,17 +40,22 @@
       </nav>
       <section class="container">
         <h1>Player example</h1>
-        <p>In general, we recommend using our <a href="https://docs.vindral.com/playout/embeddable-player">hosted player</a> to make sure it's up to date.</p>
-        <p>The bundled Player SDK is a great way to integrate inside a native app or when you want/need to use your own infrastructure.</p>
+        <p>Vindral Player is our ready-to-go web component-based player for easy integration into your web applications.</p>
 
         <div id="playback-state">Initiating...</div>
-        <div id="vindral-container"></div>
+        <vindral-player
+          id="vindral-player"
+          url="https://lb.cdn.vindral.com"
+          channel-id="vindral_demo2_ci_dbd1ca98-4e05-4a36-8215-0e8bca295ed2"
+          channel-group-id="vindral_demo_pk_932730be-db0c-46a0-a592-cfce7bdc5a43">
+        </vindral-player>
       </section>
 
       <section class="container" style="margin-top: 50px;">
         <ul>
           <li><a target="_blank" href="https://github.com/RealSprint/vindral-examples/tree/main/player">Example source code</a></li>
           <li><a target="_blank" href="https://docs.vindral.com/playout/websdk">Web SDK Documentation</a></li>
+          <li><a target="_blank" href="https://docs.vindral.com/playout/websdk/vindral-player/">Vindral Player Documentation</a></li>
         </ul>
       </section>
     </main>
