+++ vanilla-js-player-esm/index.html
@@ -28,6 +28,9 @@
       }
     </style>
 
+    <!-- Host and fetch Vindral via a CDN or your own servers -->
+    <script type="module" async src="https://cdn.jsdelivr.net/npm/@vindral/web-sdk/vindral-player-component.js"></script>
+
 </head>
   <body>
     <main class="wrapper">
@@ -38,64 +41,48 @@
       </nav>
       <section class="container">
         <h1>Vanilla JS Player example using ESM</h1>
-        <p>The bundled Player SDK is a great way to integrate inside a native app or when you want/need to use your own infrastructure.</p>
+        <p>Vindral Player is our ready-to-go web component-based player for easy integration into your web applications.</p>
 
         <div id="playback-state">Initiating...</div>
-        <div id="vindral-container"></div>
+        <vindral-player
+          id="vindral-player"
+          url="https://lb.cdn.vindral.com"
+          channel-id="vindral_demo2_ci_dbd1ca98-4e05-4a36-8215-0e8bca295ed2"
+          channel-group-id="vindral_demo_pk_932730be-db0c-46a0-a592-cfce7bdc5a43">
+        </vindral-player>
       </section>
 
       <section class="container" style="margin-top: 50px;">
         <ul>
           <li><a target="_blank" href="https://github.com/RealSprint/vindral-examples/tree/main/vanilla-js-player-esm">Example source code</a></li>
           <li><a target="_blank" href="https://docs.vindral.com/playout/websdk/">Web SDK Documentation</a></li>
+          <li><a target="_blank" href="https://docs.vindral.com/playout/websdk/vindral-player/">Vindral Player Documentation</a></li>
         </ul>
       </section>
     </main>
 
-    <!-- Host and fetch Vindral via a CDN or your own servers -->
-    <link
-      rel="stylesheet"
-      href="https://cdn.jsdelivr.net/npm/@vindral/web-sdk@3.3.4/style.css">
-    </link>
-    <script type="module">
-      import {Vindral, Player} from 'https://cdn.jsdelivr.net/npm/@vindral/web-sdk@3.3.4/+esm'
-
-      const vindralContainer = document.querySelector("#vindral-container")
+    <script>
+      const player = document.querySelector("#vindral-player")
       const playbackState = document.querySelector("#playback-state")
 
-      const player = new Player({
-        url: "https://lb.cdn.vindral.com",
-        channelId: "vindral_demo1_ci_099ee1fa-80f3-455e-aa23-3d184e93e04f",
-        // if providing a channelGroupId, the player with add the channel selector in its controls section
-        // making it possible to switch channel inline, without re-initiating the player
-        channelGroupId: "vindral_demo_pk_932730be-db0c-46a0-a592-cfce7bdc5a43",
-      }, {
-        fullscreenButtonEnabled: true,
+      player.addEventListener("vindral-instance-ready", () => {
+        // Errors are emitted when they can not be handled internally
+        // fatal errors means that the client has been unloaded and will need to be re-initialized
+        player.instance.on("error", (error) => {
+          if (error.isFatal()) {
+            // A fatal error has occurred and the instance has been unloaded, read error.message to see what
+            // This can happen if the client has been unsuccessful to connect or authentication failed
+            // In this case a new Vindral instance needs to be created to restore the session
+            console.log("fatal error: ", error.message)
+          }
+        })
+
+        // This event is emitted when the playback state changes - can be used to show a buffer spinner during buffering
+        player.instance.on("playback state", (state) => (playbackState.textContent = state))
+
+        // This event is emitted when timed metadata events occur
+        player.instance.on("metadata", (metadata) => console.log("metadata: ", metadata.content))
       })
-
-      // Errors are emitted when they can not be handled internally
-      // fatal errors means that the client has been unloaded and will need to be re-initialized
-      player.core.on("error", (error) => {
-        if (error.isFatal()) {
-          // A fatal error has occured and the instance has been unloaded, read error.message to see what
-          // This can happen if the client has been unsuccessful to connect or authentication failed
-          // In this case a new Vindral instance needs to be created to restore the session
-          console.log("fatal error: ", error.message)
-        }
-      })
-
-      // This event is emitted when the playback state changes - can be used to show a buffer spinner during buffering
-      player.core.on("playback state", (state) => (playbackState.textContent = state))
-
-      // This event is emitted when timed metadata events occur
-      player.core.on("metadata", (metadata) => console.log("metadata: ", metadata.content))
-
-      // Will connect, start the stream and try to play
-      player.core.play()
-
-      // Attaches the player view to the DOM
-      player.attach(vindralContainer)
-
     </script>
   </body>
 </html>
